More work on optimizers: Seems to be fine now
