+++ test/test_backward_propagations.cpp
@@ -1,15 +1,23 @@
+#include "activation_functions.hpp"
 #include "activation_layer.hpp"
 #include "defines.hpp"
-#include "robbie.hpp"
+#include "dropout_layer.hpp"
+#include "fc_layer.hpp"
+#include "optimizers.hpp"
 #include <fmt/format.h>
 #include <fmt/ostream.h>
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
 #include <cstddef>
+#include <memory>
 
 template<typename scalar>
 void test_backward_propagation( Robbie::Layer<scalar> * layer, const Robbie::Matrix<scalar> & x0, size_t output_size )
 {
+    // Use the DoNothing optimizer
+    // Otherwise the weights will change on the backward propagation and we cannot compare to subsequent forward optimizations
+    layer->opt = std::make_unique<Robbie::Optimizers::DoNothing<scalar>>();
+
     // The loss function is the sum of outputs E = y1 + ... + yN
     auto loss0 = layer->forward_propagation( x0 ).colwise().sum().eval();
 
@@ -36,6 +44,7 @@ void test_backward_propagation( Robbie::Layer<scalar> * layer, const Robbie::Mat
 
         auto derivative_fd = ( loss_new - loss0 ) / epsilon;
 
+        fmt::print( "output_error[{}] = {}\n", i, fmt::streamed( output_error.row( i ) ) );
         fmt::print( "input_error[{}] = {}\n", i, fmt::streamed( input_error.row( i ) ) );
         fmt::print( "derivative_fd = {}\n", fmt::streamed( derivative_fd ) );
